+++ packages/lab/src/__tests__/__e2e__/text/Text.cy.tsx
