+++ packages/lab/src/contact-details/ContactMetadataItem.tsx
@@ -1,7 +1,7 @@
 import { makePrefixer } from "@jpmorganchase/uitk-core";
 import { IconProps } from "@jpmorganchase/uitk-icons";
 import { ComponentType, forwardRef, HTMLAttributes } from "react";
-import { Div } from "../typography";
+import { Div } from "../text";
 import { ValueComponentProps } from "./internal";
 import { MailLinkComponent } from "./";
