+++ packages/lab/src/contact-details/ContactSecondaryInfo.tsx
@@ -2,7 +2,7 @@ import { makePrefixer, useId } from "@jpmorganchase/uitk-core";
 import { IconProps } from "@jpmorganchase/uitk-icons";
 import cn from "classnames";
 import { ComponentType, forwardRef, HTMLAttributes, useEffect } from "react";
-import { Div } from "../typography";
+import { Div } from "../text";
 import { useContactDetailsContext, ValueComponentProps } from "./internal";
 
 const withBaseName = makePrefixer("uitkContactSecondaryInfo");
