+++ packages/lab/src/index.ts
@@ -50,7 +50,7 @@ export * from "./toggle-button";
 export * from "./toolbar";
 export * from "./tokenized-input";
 export * from "./tree";
-export * from "./typography";
+export * from "./text";
 export * from "./window";
 export * from "./utils";
