+++ packages/lab/src/text/Figure.tsx
