+++ packages/lab/src/text/LabelCaption.tsx
