+++ packages/lab/src/text/Span.tsx
