+++ packages/lab/stories/text/Table/Table.css
